requirements.txt
colorama
transformers
openai
Pillow
tabulate
PyYAML
pandas
